+++ assembler/src/common/assembly_graph/core/observable_graph.hpp
@@ -150,7 +150,7 @@ public:
 
     EdgeId MergePath(const std::vector<EdgeId> &path,
                      bool safe_merging = true,
-                     std::vector<uint32_t> overlaps = std::vector<uint32_t>());
+                     std::vector<size_t> overlaps = std::vector<size_t>());
 
     std::pair<EdgeId, EdgeId> SplitEdge(EdgeId edge, size_t position);
 
@@ -446,7 +446,7 @@ template<class DataMaster>
 typename ObservableGraph<DataMaster>::EdgeId
         ObservableGraph<DataMaster>::MergePath(const std::vector<EdgeId> &path,
                                                bool safe_merging,
-                                               std::vector<uint32_t> overlaps) {
+                                               std::vector<size_t> overlaps) {
     VERIFY(!path.empty());
     for (size_t i = 0; i < path.size(); i++)
         for (size_t j = i + 1; j < path.size(); j++) {
@@ -462,12 +462,12 @@ typename ObservableGraph<DataMaster>::EdgeId
     VertexId v1 = base::EdgeStart(corrected_path[0]);
     VertexId v2 = base::EdgeEnd(corrected_path[corrected_path.size() - 1]);
     std::vector<const EdgeData *> to_merge;
-    std::vector<uint32_t> local_overlaps;
+    std::vector<size_t> local_overlaps;
     for (auto it1 = corrected_path.begin(), it2 = std::next(it1); it2 != corrected_path.end(); ++it1, ++it2) {
         to_merge.push_back(&(base::data(*it1)));
         VertexId end = base::EdgeEnd(*it1);
         VERIFY(end == base::EdgeStart(*it2));
-        uint32_t overlap = base::data(end).overlap();
+        size_t overlap = base::data(end).overlap();
         local_overlaps.push_back(overlap);
     }
     if (not overlaps.empty()) {
